+++ ETLBoxTest/src/DataFlow/TestDataFlowDBSource.cs
@@ -54,7 +54,7 @@ namespace ALE.ETLBoxTest {
             source.LinkTo(dest);
             source.Execute();
             dest.Wait();
-            Assert.AreEqual(3, SqlTask.ExecuteScalar<int>("Check destination table", "select count(*) from test.Destination"));
+            Assert.AreEqual(3, RowCountTask.Count("test.Destination"));
         }
 
         /*
@@ -74,7 +74,7 @@ namespace ALE.ETLBoxTest {
             source.LinkTo(dest);
             source.Execute();
             dest.Wait();
-            Assert.AreEqual(3, SqlTask.ExecuteScalar<int>("Check destination table", "select count(*) from test.Destination"));
+            Assert.AreEqual(3, RowCountTask.Count("test.Destination"));
         }
 
         /*
@@ -94,7 +94,7 @@ namespace ALE.ETLBoxTest {
             source.LinkTo(dest);
             source.Execute();
             dest.Wait();
-            Assert.AreEqual(3, SqlTask.ExecuteScalar<int>("Check destination table", "select count(*) from test.Destination"));
+            Assert.AreEqual(3, RowCountTask.Count("test.Destination"));
         }
 
         /*
@@ -115,7 +115,7 @@ namespace ALE.ETLBoxTest {
             source.LinkTo(dest);
             source.Execute();
             dest.Wait();
-            Assert.AreEqual(3, SqlTask.ExecuteScalar<int>("Check destination table", "select count(*) from test.Destination"));
+            Assert.AreEqual(3, RowCountTask.Count("test.Destination"));
         }